\name{alr}
\alias{alr}
\title{Additive log ratio (alr) transformation for compositions}
\usage{
alr(x)
}
\arguments{
  \item{x}{a compositional vector or dataframe of row-wise
  compositions}
}
\value{
an alr transformed vector or dataframe
}
\description{
Implements the alr transformation for vectors and
dataframes
}

