\name{clr}
\alias{clr}
\title{Centered log ratio (clr) transformation for compositions}
\usage{
clr(x)
}
\arguments{
  \item{x}{a compositional vector or dataframe of row-wise
  compositions}
}
\value{
an clr transformed vector or dataframe
}
\description{
Implements the clr transformation for vectors and
dataframes
}

