\name{FASTIN}
\alias{FASTIN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gui for the FASTIN analysis
}
\description{
Provides a gui to input data to be analysed with the
\code{\link{run_MCMC}} function. The gui rolls the functionality of most
of the functions contained within FASTIN-R into a single interface, but
some options (plottin options mainly) are not available through the gui.
}
\usage{
FASTIN()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Please consult the help individual functions to obtain help on a
  particular topic
}
\references{
Neubauer.P. and Jensen, O.P. (in prep)
}
\author{
Philipp Neubauer
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{addFA}},\code{\link{addSI}},\code{\link{run_MCMC}},\code{\link{diags}}
}

